Fix real compile errors surfaced by CI
CI ran for the first time and confirmed several mistakes made without
access to a Mac compiler:

- SceneKit's SCNVector3/SCNMatrix4 fields are CGFloat on this SDK, not
  Float as I'd assumed -- reverted the Float conversions in
  SkinModelBuilder's UV transform, the rig-building positions, the
  pet's scale, and the cursor-follow head tracking.
- NSAnimationContext's method is runAnimationGroup(_:completionHandler:),
  not runAnimationContext -- fixed all three call sites.
- SCNAction.scaleBy only scales uniformly (no per-axis x/y/z variant);
  rewrote the landing squash-and-stretch as a customAction that
  manually interpolates a non-uniform SCNVector3 scale instead.
- SkinTextureLoader.resample had two parameters both labeled `to:`,
  which requires `to:to:` at the call site rather than a single
  positional argument -- renamed to width:/height:.
- Quieted an always-succeeds conditional cast warning in BatteryMonitor.

+++ Sources/BoxBucko/AppController.swift
@@ -331,8 +331,8 @@ final class AppController: NSObject {
             let maxDistance: CGFloat = 500
             let clampedDx = max(-maxDistance, min(maxDistance, dx))
             let clampedDy = max(-maxDistance, min(maxDistance, dy))
-            let yaw = Float(max(-0.7, min(0.7, clampedDx / 220)))
-            let pitch = Float(max(-0.5, min(0.5, -clampedDy / 260)))
+            let yaw = max(-0.7, min(0.7, clampedDx / 220))
+            let pitch = max(-0.5, min(0.5, -clampedDy / 260))
             rig.head.eulerAngles.y = rig.head.eulerAngles.y * 0.7 + yaw * 0.3
             rig.head.eulerAngles.x = rig.head.eulerAngles.x * 0.7 + pitch * 0.3
         }
